requirements.txt
selenium==4.1.3
chromedriver_autoinstaller==0.3.1
python-docx==0.8.11
urllib3
requests




